Move ConnectionManager to own file
